Make some changes to wishlist menu js
I've made some changes to wishlist menu js, so now product id attribute
adds not to product card, it adds directly to button, which is more
unified and easy to work with. I've also added delete button to every
product card in wishlist menu, so now it can be deleted directly from
menu. I've also added circle with amount of products in wishlist menu to
mobile nav wishlist button. And one of the most important things is that
all addToWishlistButtons are managed from SetupWishlistMenu class, so
now it doesn't matter on which page button, it's just need to have
proper attribute. I've also made synchronization between all buttons, so
if user deletes product by using delete button in menu,
active class of addToWishlistButton with the same id will be removed.
And it also concerns buttons with the same id in all open pages.